backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.0.1
torchvision==0.15.2
numpy<2.0
timm==0.9.10
Pillow==10.1.0
huggingface-hub==0.19.4
pydantic==2.5.2
librosa==0.10.1
soundfile==0.12.1
transformers==4.35.2
accelerate==0.25.0
peft==0.6.2
pydub==0.25.1
av==11.0.0
mutagen==1.47.0